moved all references to config and trigonometry to utils folder and fixed doc references
