+++ src/PanCAD/svg/file.py
@@ -6,8 +6,8 @@ import xml.etree.ElementTree as ET
 
 from PanCAD.svg import element_utils as seu
 from PanCAD.svg.elements import SVGElement
-from PanCAD import file_handlers
-from PanCAD.file_handlers import InvalidAccessModeError
+from PanCAD.utils import file_handlers
+from PanCAD.utils.file_handlers import InvalidAccessModeError
 
 class SVGFile(ET.ElementTree):
     """A class for svg files with a single svg element at the top that
