+++ tests/test_freecad_object_wrappers.py
@@ -5,7 +5,7 @@ import unittest
 sys.path.append('src')
 
 from PanCAD.freecad import object_wrappers as fcow
-from PanCAD import file_handlers
+from PanCAD.utils import file_handlers
 
 import FreeCAD as App
 import Part
